+++ app/src/main/java/org/xwiki/android/sync/activities/OIDC/OIDCActivity.kt
@@ -1,22 +1,54 @@
 package org.xwiki.android.sync.activities.OIDC
 
 import android.accounts.Account
-import android.accounts.AccountAuthenticatorActivity
 import android.accounts.AccountManager
 import android.app.Activity
 import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.util.Log
 import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
+import com.google.api.client.auth.oauth2.AuthorizationCodeFlow
+import com.google.api.client.auth.oauth2.BearerToken
+import com.google.api.client.auth.oauth2.ClientParametersAuthentication
+import com.google.api.client.auth.openidconnect.IdTokenResponse
+import com.google.api.client.http.GenericUrl
+import com.google.api.client.http.javanet.NetHttpTransport
+import com.google.api.client.json.jackson2.JacksonFactory
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import org.xwiki.android.sync.*
 import org.xwiki.android.sync.activities.OIDC.OIDCActivity.OIDCActivity.selectedAc
 import org.xwiki.android.sync.auth.AuthenticatorActivity
 import org.xwiki.android.sync.databinding.ActOidcChooseAccountBinding
 import org.xwiki.android.sync.utils.AccountClickListener
-import org.xwiki.android.sync.utils.getValue
-import org.xwiki.android.sync.utils.putValue
+import org.xwiki.android.sync.utils.extensions.TAG
 
-class OIDCActivity: AccountAuthenticatorActivity(), AccountClickListener {
+private fun createAuthorizationCodeFlow(): AuthorizationCodeFlow {
+    return AuthorizationCodeFlow.Builder(
+        BearerToken.authorizationHeaderAccessMethod(),
+        NetHttpTransport(),
+        JacksonFactory(),
+        GenericUrl(TOKEN_SERVER_URL),
+        ClientParametersAuthentication(
+            selectedAc,
+            ""
+        ),
+        selectedAc,
+        AUTHORIZATION_SERVER_URL
+    ).apply {
+        scopes = mutableListOf(
+            "openid",
+            "offline_access",
+            "profile"
+        )
+    }.build()
+}
+
+class OIDCActivity: AppCompatActivity(), AccountClickListener {
 
     private lateinit var binding: ActOidcChooseAccountBinding
 
@@ -30,10 +62,6 @@ class OIDCActivity: AccountAuthenticatorActivity(), AccountClickListener {
         binding = DataBindingUtil.setContentView(this, R.layout.act_oidc_choose_account)
 
         init()
-
-        binding.lvAddAnotherAccount.setOnClickListener {
-            addNewAccount()
-        }
     }
 
     private fun init() {
@@ -45,24 +73,46 @@ class OIDCActivity: AccountAuthenticatorActivity(), AccountClickListener {
         }
 
         val adapter = OIDCAccountAdapter(this, availableAccountsList, this)
+
+        binding.lvAddAnotherAccount.setOnClickListener {
+            addNewAccount()
+        }
+
         binding.lvSelectAccount.adapter = adapter
+    }
+
+    private fun extractAuthorizationCode(intent: Intent): String? {
+        val data = intent.dataString
+        val uri = Uri.parse(data)
+
+        return uri.getQueryParameter("code")
+    }
+
+    private fun requestAccessToken(flow: AuthorizationCodeFlow, authorizationCode: String?) {
+        appCoroutineScope.launch {
+            try {
+                val idTokenResponse = IdTokenResponse.execute(
+                    flow.newTokenRequest(
+                        authorizationCode
+                    ).setRedirectUri(
+                        REDIRECT_URI
+                    )
+                )
+                Log.d(TAG, idTokenResponse.parseIdToken().payload["sub"].toString())
+                val accessToken = idTokenResponse.accessToken
+
+                withContext(Dispatchers.Main) { // launch on UI thread. TODO:: Check necessarily of UI scope
+                    sendResult(accessToken)
+                }
+            } catch (ex: Exception) {
+                Log.e(TAG, ex.message)
+            }
+        }
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
         super.onActivityResult(requestCode, resultCode, data)
         when (requestCode) {
-            REQUEST_ACCESS_TOKEN -> {
-                val accessToken = getValue(this, "ACCESS_TOKEN", "")
-                if (accessToken.isNullOrEmpty()) {
-                    Toast.makeText(this, "Something went wrong. Please try again.", Toast.LENGTH_SHORT).show()
-                } else {
-                    val i = Intent()
-                    i.putExtra(AccountManager.KEY_AUTHTOKEN, accessToken)
-                    setResult(Activity.RESULT_OK, i)
-                    finish()
-                }
-                putValue(this, "ACCESS_TOKEN", "")
-            }
             REQUEST_NEW_ACCOUNT -> {
                 val mAccountManager = AccountManager.get(applicationContext)
                 val availableAccountsList = mAccountManager.getAccountsByType(ACCOUNT_TYPE)
@@ -71,14 +121,44 @@ class OIDCActivity: AccountAuthenticatorActivity(), AccountClickListener {
                 binding.lvSelectAccount.adapter = adapter
 
                 selectedAc = data?.getStringExtra(AccountManager.KEY_ACCOUNT_NAME).toString()
-                startActivityForResult(Intent (this, OIDCAuthenticatorActivity::class.java), REQUEST_ACCESS_TOKEN)
+                startAuthorization()
             }
         }
     }
 
+    private fun sendResult(accessToken: String?) {
+        if (accessToken.isNullOrEmpty()) {
+            Toast.makeText(this, "Something went wrong. Please try again.", Toast.LENGTH_SHORT).show()
+        } else {
+            val i = Intent()
+            i.putExtra(AccountManager.KEY_AUTHTOKEN, accessToken)
+            setResult(Activity.RESULT_OK, i)
+            finish()
+        }
+    }
+
+    override fun onNewIntent(intent: Intent?) {
+        super.onNewIntent(intent)
+        intent ?: return
+        val authorizationCode = extractAuthorizationCode(intent)
+        requestAccessToken(createAuthorizationCodeFlow(), authorizationCode)
+    }
+
     override fun invoke(selectedAcc: Account) {
         selectedAc = selectedAcc.name
-        startActivityForResult(Intent (this, OIDCAuthenticatorActivity::class.java), REQUEST_ACCESS_TOKEN)
+        startAuthorization()
+    }
+
+    private fun startAuthorization () {
+        try {
+            val flow: AuthorizationCodeFlow = createAuthorizationCodeFlow()
+            flow.newAuthorizationUrl().setRedirectUri(REDIRECT_URI).build().let {
+                val browserIntent = Intent(Intent.ACTION_VIEW, Uri.parse(it))
+                startActivity(browserIntent)
+            }
+        } catch (ex: Exception) {
+            Log.e(TAG, "Failed")
+        }
     }
 
     private fun addNewAccount() {
